Toggle to show Replay Box
In the style options, there will be a toggle to show the replay box for cards. Even if enabled, the green box will only be visible if there is text in it. Due to the nature of card style options, when switching back to the card list, it will look like the setting has been reverted, but switching to another card and then back will refresh it.